% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ops.R
\name{pak_intersect}
\alias{pak_intersect}
\title{Intersect two sf objects}
\usage{
pak_intersect(x, y)
}
\arguments{
\item{x}{An sf object.}

\item{y}{An sf object.}
}
\value{
An sf object of the intersection in the CRS of x.
}
\description{
Returns the geometric intersection of two sf objects with CRS
alignment handled automatically.
}
