% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{pk_version}
\alias{pk_version}
\title{Check package version and update status}
\usage{
pk_version(quiet = FALSE)
}
\arguments{
\item{quiet}{Logical. If FALSE (default), prints to console.}
}
\value{
Invisibly, a list with installed and latest versions.
}
\description{
Check package version and update status
}
\examples{
\donttest{
pk_version()
}
}
