% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{pkmapr-deprecated}
\alias{pkmapr-deprecated}
\alias{pak_crs_suggest}
\alias{pak_dictionary}
\alias{pak_join}
\alias{pak_area}
\alias{pak_distance}
\alias{pak_centroid}
\alias{pak_bbox}
\alias{pak_project}
\alias{pak_neighbors}
\alias{pak_points_in}
\alias{pak_buffer}
\alias{pak_union}
\alias{pak_intersect}
\alias{pak_map}
\alias{pak_map_interactive}
\alias{pak_basemap}
\title{Deprecated functions in pkmapr}
\usage{
pak_crs_suggest(...)

pak_dictionary(...)

pak_join(...)

pak_area(...)

pak_distance(...)

pak_centroid(...)

pak_bbox(...)

pak_project(...)

pak_neighbors(...)

pak_points_in(...)

pak_buffer(...)

pak_union(...)

pak_intersect(...)

pak_map(...)

pak_map_interactive(...)

pak_basemap(...)
}
\value{
No return value, called for side effects (deprecation warnings).
}
\description{
These functions were deprecated.
Please use the new \code{pk_*} functions instead.
}
\keyword{internal}
