% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkmapr-package.R
\docType{package}
\name{pkmapr-package}
\alias{pkmapr}
\alias{pkmapr-package}
\title{pkmapr: Pakistan Administrative Boundaries Toolkit}
\description{
Provides a tidy interface to Pakistan's official administrative boundary data from the United Nations Office for the Coordination of Humanitarian Affairs (OCHA). Downloads and caches spatial data at country, province, district, and tehsil levels as sf objects compatible with the tidyverse and geospatial ecosystem. Includes utilities for geographic dictionary lookup, coordinate reference system selection, spatial measurement, and neighbour structure construction for use with spdep, ggplot2, leaflet, and related packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://abdullahumer1101.github.io/pkmapr}
  \item \url{https://github.com/abdullahumer1101/pkmapr}
  \item Report bugs at \url{https://github.com/abdullahumer1101/pkmapr/issues}
}

}
\author{
\strong{Maintainer}: Abdullah Umer \email{abdullahumer1101@gmail.com} (\href{https://orcid.org/0009-0008-4082-8394}{ORCID})

}
\keyword{internal}
