% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_management.R
\name{pkmapr_clear_cache}
\alias{pkmapr_clear_cache}
\title{Clear the pkmapr cache}
\usage{
pkmapr_clear_cache()
}
\value{
NULL, invisibly.
}
\description{
Deletes all locally cached files. They will be re-downloaded on the
next function call.
}
